chore: move `grass.cpp` into `src/renderer/grass.cpp`
